+++ BankAccount/Modules/Main/RegisterCell/RegisterCell.swift
@@ -1,0 +1,16 @@
+//
+//  RegisterCell.swift
+//  BankAccount
+//
+//  Created by Egor Syrtcov on 4/3/20.
+//  Copyright © 2020 Egor Syrtcov. All rights reserved.
+//
+
+import Foundation
+
+enum RegisterCell {
+    static let customCell = "CustomTableViewCell"
+    static let currentCell = "CollectionInTableViewCell"
+    static let collectionViewCell = "CollectionViewCell"
+    static let lastCollectionCell = "LastCollectionCell"
+}
