+++ BankAccount/Modules/Main/Service/Service.swift
@@ -29,9 +29,9 @@ class Service {
         urlString = getStringUrl(forKey: KeyUrl.adressParse.rawValue)
     }
     
-    func fetchRequestBillingItems(path: AdressBankAccount, completion: @escaping ([Billing]?) ->()) {
+    func fetchRequestBillingItems(completion: @escaping ([Billing]?) ->()) {
         
-        Alamofire.request("\(urlString)\(path.rawValue)").responseArray { (response: DataResponse<[Billing]>) in
+        Alamofire.request("\(urlString)\(AdressBankAccount.allBilling)").responseArray { (response: DataResponse<[Billing]>) in
             var billingItems = [Billing]()
             let billingArray = response.result.value
             
@@ -42,8 +42,8 @@ class Service {
         }
     }
     
-    func fetchRequestTransactionItems(path: AdressBankAccount, completion: @escaping ([Transaction]?) ->()) {
-        Alamofire.request("\(urlString)\(path.rawValue)").responseArray { (response: DataResponse<[Transaction]>) in
+    func fetchRequestTransactionItems(completion: @escaping ([Transaction]?) ->()) {
+        Alamofire.request("\(urlString)\(AdressBankAccount.allTransaction)").responseArray { (response: DataResponse<[Transaction]>) in
             var transactionItems = [Transaction]()
             let transactionArray = response.result.value
